Split backend http services
BE:FE: Update API routes
