+++ backend/src/main/scala/finalmp/models/events/decoder/GameEventDec.scala
@@ -18,7 +18,7 @@ object GameEventDec {
           case "player.connect" => Decoder[PlayerConnect].apply(c)
           case "player.disconnect" => Decoder[PlayerDisconnect].apply(c)
           case "player.update" => Decoder[PlayerUpdate].apply(c)
-          case _ => Left(DecodingFailure(s"Unknown game event message ${messageType}", CursorOp.DownField("type") :: Nil))
+          case _ => Left(DecodingFailure(s"Unknown game event message $messageType", CursorOp.DownField("type") :: Nil))
         }
       } yield out
     }
